# openapi-specification.yaml
openapi: 3.1.0
info:
  title: ProcGrid B2B E-commerce Platform API
  description: |
    Production-grade B2B e-commerce platform for producer-buyer agricultural product trading.
    
    ## Authentication
    All protected endpoints require JWT Bearer token obtained from Keycloak authentication server.
    
    ## Rate Limiting
    - Public endpoints: 100 requests/minute
    - Authenticated endpoints: 1000 requests/minute
    
    ## Error Handling
    All errors follow RFC 7807 Problem Details for HTTP APIs standard.
  version: 1.0.0
  contact:
    name: ProcGrid API Support
    email: api-support@procgrid.com
    url: https://procgrid.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://procgrid.com/terms
servers:
  - url: https://api.procgrid.com/v1
    description: Production server
  - url: https://staging-api.procgrid.com/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Local development server
security:
  - bearerAuth: []
paths:
...
  /products:
    post:
      tags:
        - Product Management
      summary: Create product lot
      description: Create new product lot (Producer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '201':
          description: Product lot created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/ValidationError'

    get:
      tags:
        - Product Management
      summary: List products
      description: Get paginated list of product lots
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortDirectionParam'
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category name
        - name: location
          in: query
          schema:
            type: string
          description: Filter by location
        - name: minPrice
          in: query
          schema:
            type: number
            format: double
          description: Minimum price filter
        - name: maxPrice
          in: query
          schema:
            type: number
            format: double
          description: Maximum price filter
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProductPageResponse'
  /products/{id}:
    get:
      tags:
        - Product Management
      summary: Get product details
      description: Retrieve detailed product information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "prod_12345"
      responses:
        '200':
          description: Product details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      tags:
        - Product Management
      summary: Update product
      description: Update product information (Producer only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /products/search:
    get:
      tags:
        - Product Management
      summary: Search products
      description: Advanced product search with filters
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query
        - name: category
          in: query
          schema:
            type: string
          description: Category filter
        - name: priceRange
          in: query
          schema:
            type: string
            pattern: '^\\d+(-\\d+)?$'
          description: Price range (e.g., "100-500")
        - name: location
          in: query
          schema:
            type: string
          description: Location filter
        - name: qualityGrade
          in: query
          schema:
            $ref: '#/components/schemas/QualityGrade'
          description: Quality grade filter
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Search results retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProductPageResponse'

  # ============ RFQ / QUOTE SERVICE ============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from Keycloak authentication
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Page number (0-based)
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
    SortByParam:
      name: sortBy
      in: query
      schema:
        type: string
        default: "id"
      description: Field to sort by
    SortDirectionParam:
      name: sortDirection
      in: query
      schema:
        type: string
        enum: [ASC, DESC]
        default: "ASC"
      description: Sort direction
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
        - message
        - timestamp
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Human-readable message
        errorCode:
          type: string
          description: Error code for failed requests
          nullable: true
        data:
          description: Response data
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
    ErrorResponse:
      type: object
      required:
        - success
        - message
        - errorCode
        - timestamp
        - path
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Error description
        errorCode:
          type: string
          description: Specific error code
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: API endpoint path

    # ============ AUTHENTICATION SCHEMAS ============
...
    QualityGrade:
      type: string
      enum:
        - PREMIUM
        - GRADE_A
        - GRADE_B
        - STANDARD
      description: Product quality grade
    ProductUnit:
      type: string
      enum:
        - KG
        - TONNE
        - QUINTAL
        - PIECE
        - LITER
      description: Product measurement unit
    ProductStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - OUT_OF_STOCK
        - DISCONTINUED
      description: Product availability status
...
    Product:
      type: object
      required:
        - productId
        - producerId
        - categoryId
        - title
        - description
        - pricePerUnit
        - quantity
        - unit
        - location
        - qualityGrade
        - status
        - createdAt
      properties:
        productId:
          type: string
        producerId:
          type: string
        categoryId:
          type: string
        title:
          type: string
        description:
          type: string
        pricePerUnit:
          type: number
          format: double
          minimum: 0
        quantity:
          type: number
          format: double
          minimum: 0
        unit:
          $ref: '#/components/schemas/ProductUnit'
        location:
          type: string
        qualityGrade:
          $ref: '#/components/schemas/QualityGrade'
        harvestDate:
          type: string
          format: date
          nullable: true
        expiryDate:
          type: string
          format: date
          nullable: true
        certifications:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
            format: uri
        status:
          $ref: '#/components/schemas/ProductStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProductCreateRequest:
      type: object
      required:
        - categoryId
        - title
        - description
        - pricePerUnit
        - quantity
        - unit
        - location
        - qualityGrade
      properties:
        categoryId:
          type: string
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        pricePerUnit:
          type: number
          format: double
          minimum: 0
        quantity:
          type: number
          format: double
          minimum: 0.01
        unit:
          $ref: '#/components/schemas/ProductUnit'
        location:
          type: string
          maxLength: 500
        qualityGrade:
          $ref: '#/components/schemas/QualityGrade'
        harvestDate:
          type: string
          format: date
          nullable: true
        expiryDate:
          type: string
          format: date
          nullable: true
        certifications:
          type: array
          items:
            type: string
    ProductUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        pricePerUnit:
          type: number
          format: double
          minimum: 0
        quantity:
          type: number
          format: double
          minimum: 0
        location:
          type: string
          maxLength: 500
        status:
          $ref: '#/components/schemas/ProductStatus'
    ProductPageResponse:
      allOf:
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Product'
        - $ref: '#/components/schemas/PageInfo'
    PageInfo:
      type: object
      required:
        - page
        - size
        - totalElements
        - totalPages
        - first
        - last
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        first:
          type: boolean
        last:
          type: boolean
        sortBy:
          type: string
        sortDirection:
          type: string
          enum: [ASC, DESC]

    # ============ QUOTE SCHEMAS ============
...
  responses:
...
    ForbiddenError:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Insufficient permissions"
            errorCode: "AUTH_004"
            timestamp: "2025-11-15T10:30:00Z"
            path: "/api/v1/products/123"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Resource not found"
            errorCode: "USER_001"
            timestamp: "2025-11-15T10:30:00Z"
            path: "/api/v1/users/999"
...
    ValidationError:
      description: Validation error - Invalid input data
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  data:
                    type: object
                    additionalProperties:
                      type: string
          example:
            success: false
            message: "Validation failed"
            errorCode: "VAL_001"
            data:
              email: "Invalid email format"
              password: "Password must be at least 6 characters"
            timestamp: "2025-11-15T10:30:00Z"
            path: "/api/v1/register/buyer"
...